// src/app/routes/project/interfaces/panel.interface.ts
/*
 * Copyright 2021 Google LLC
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     https://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { IPanel, PropertyPanelState } from 'cd-interfaces';
import { IActivityConfig } from './activity.interface';
import { IPropertyModeConfig } from './properties.interface';

export interface IPanelsState {
  currentActivity: IActivityConfig;
  propertyMode: IPropertyModeConfig;
  searchString?: string;
  componentPanelActiveTabIndex?: number;
  leftPanel: IPanel;
  rightPanel: IPanel;
  bottomPanel: IPanel;
  symbolMode: boolean;
  isolatedSymbolId?: string;
  propertiesPanelState?: PropertyPanelState;
  recordStateChanges: boolean;
  keyboardShortcutsModalOpen: boolean;
}

export interface IPanelHeaderAction {
  readonly title: string;
  readonly action?: string;
}